AF-7: Separated Admin from Main modules, deleted Search due to changed requirements
