+++ afisha-client/src/app/app.component.html
@@ -1,4 +1,3 @@
 <!--The content below is only a placeholder and can be replaced.-->
-<app-navbar *ngIf="!show()"></app-navbar>
 <router-outlet></router-outlet>
 <!-- here goes footer -->
