+++ afisha-client/src/app/app.component.ts
@@ -1,6 +1,4 @@
 import { Component } from '@angular/core';
-import {PATH} from "./shared/constants/path.constant";
-import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-root',
@@ -10,16 +8,4 @@ import {Router} from "@angular/router";
 export class AppComponent {
   title = 'app';
 
-  constructor(private router: Router) {
-
-  }
-
-  /**
-   * This function will check if the current path is /search, if true it hides the navbar
-   * @returns {boolean}
-   */
-  show(): boolean {
-    let bool = this.router.url === "/" + PATH.SEARCH || this.router.url === "/" + PATH.ADMIN;
-    return bool;
-  }
 }
