+++ afisha-client/src/app/app.module.ts
@@ -8,13 +8,8 @@ import {RouterModule} from "@angular/router";
 import {FlexLayoutModule} from "@angular/flex-layout";
 import {Error404Component} from './error/error404.component';
 import {appRoutes} from "./app.routes";
-import {NavbarComponent} from './navbar/navbar.component';
-import {EventsListComponent} from './events/events-list.component';
-import {EventDetailsComponent} from './events/event-details.component';
 import {EventService} from "./shared/services/event.service";
-import {EventComponent} from './events/event.component';
 import {GuardEventRouteService} from "./shared/guards/guard-event-route.service";
-import {SearchComponent} from './search/search.component';
 import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
 import {AuthInterceptorService} from "./shared/services/auth-interceptor.service";
 import {AuthService} from "./shared/services/auth.service";
@@ -23,12 +18,7 @@ import {EventResolverService} from "./shared/services/event-resolver.service";
 @NgModule({
   declarations: [
     AppComponent,
-    Error404Component,
-    NavbarComponent,
-    EventsListComponent,
-    EventDetailsComponent,
-    EventComponent,
-    SearchComponent
+    Error404Component
   ],
   imports: [
     BrowserModule,
