--- afisha-client/src/app/events/event-details.component.html
@@ -1,17 +1,0 @@
-<mat-card [routerLink] = "['/',path.EVENTS, event._id]" class="example-card">
-  <mat-card-header>
-    <div mat-card-avatar class="example-header-image"></div>
-    <mat-card-title>{{event.title}}</mat-card-title>
-    <mat-card-subtitle>{{event.abstract}}</mat-card-subtitle>
-  </mat-card-header>
-  <img mat-card-image src="" alt="Photo of a Shiba Inu">
-  <mat-card-content>
-    <p>
-      {{event.description}}
-    </p>
-  </mat-card-content>
-  <mat-card-actions>
-    <button mat-button>LIKE</button>
-    <button mat-button>SHARE</button>
-  </mat-card-actions>
-</mat-card>
