--- afisha-client/src/app/events/event-details.component.scss
@@ -1,5 +1,0 @@
-.example-card {
-  width: 100%;
-  height: 100%;
-}
-
