+++ afisha-client/src/app/main/events/event-details.component.scss
@@ -1,0 +1,5 @@
+.example-card {
+  width: 100%;
+  height: 100%;
+}
+
