+++ afisha-client/src/app/main/events/event.component.html
@@ -1,0 +1,6 @@
+<p>{{event.title}}</p>
+<p>{{event.abstract}}</p>
+<img mat-card-image src="{{event.imageUrl}}" alt="Photo of a Shiba Inu">
+<p>
+  {{event.description}}
+</p>
