+++ afisha-client/src/app/main/events/events-list.component.html
@@ -1,0 +1,6 @@
+<div class="flex-container">
+  <div class="flex-item" *ngFor="let event of events">
+    <app-event-details [event] = (event)></app-event-details>
+  </div>
+</div>
+
