+++ afisha-client/src/app/main/events/events-list.component.scss
@@ -1,0 +1,18 @@
+.flex-container {
+  display: flex;
+  flex-direction: row;
+  flex-flow: row wrap;
+  justify-content: space-around;
+}
+
+@media all and (max-width: 800px) {
+  .flex-container {
+    justify-content: flex-start;
+  }
+}
+
+.flex-item {
+  height: auto;
+  max-width: 300px;
+  margin-top: 30px;
+}
