+++ afisha-client/src/app/main/main.component.html
@@ -1,0 +1,2 @@
+<app-navbar></app-navbar>
+<router-outlet></router-outlet>
