+++ afisha-client/src/app/main/navbar/navbar.component.scss
