+++ afisha-client/src/app/main/user/login.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {Router} from "@angular/router";
-import {PATH} from "../shared/constants/path.constant";
-import {AuthService} from "../shared/services/auth.service";
-import {IUser} from "../shared/models/user.model";
+import {PATH} from "../../shared/constants/path.constant";
+import {AuthService} from "../../shared/services/auth.service";
+import {IUser} from "../../shared/models/user.model";
 import {HttpErrorResponse} from "@angular/common/http";
 
 @Component({
@@ -30,7 +30,7 @@ export class LoginComponent implements OnInit {
    * This function redirects to the default path /search
    */
   cancel(): void{
-    this.router.navigate([PATH.SEARCH]);
+    this.router.navigate([PATH.ROOT]);
   }
 
   login(loginValues: IUser): void{
