+++ afisha-client/src/app/main/user/register.component.scss
