+++ afisha-client/src/app/main/user/user.module.ts
@@ -5,7 +5,7 @@ import {userRoutes} from "./user.routes";
 import { LoginComponent } from './login.component';
 import {ReactiveFormsModule} from "@angular/forms";
 import {RegisterComponent} from "./register.component";
-import {MaterialModule} from "../shared/material.module";
+import {MaterialModule} from "../../shared/material.module";
 
 @NgModule({
   imports: [
