+++ afisha-client/src/app/main/user/user.routes.ts
@@ -1,5 +1,5 @@
 import {LoginComponent} from "./login.component";
-import {PATH} from "../shared/constants/path.constant";
+import {PATH} from "../../shared/constants/path.constant";
 import {RegisterComponent} from "./register.component";
 
 export const userRoutes = [
