--- afisha-client/src/app/search/search.component.html
@@ -1,58 +1,0 @@
-<div class="search-body">
-  <div class="container">
-    <nav>
-      <img src="../../assets/logo.png" alt="Omnifood logo" class="logo">
-          <ul class="search-nav">
-            <li><a [routerLink] = "['/', path.ABOUT]">About us</a></li>
-            <li><a [routerLink] = "['/', path.EVENTS]">List of events</a></li>
-            <li><a [routerLink] = "['/', path.USER, path.LOGIN]">Member</a></li>
-          </ul>
-    </nav>
-  </div>
-  <div class="container search-container">
-    <div class="row">
-      <div class="col-lg-12">
-        <div class="content">
-          <h1>FIND THE BEST EVENT FOR YOU</h1>
-          <h3></h3>
-          <hr>
-          <div class="card border-primary container">
-            <div class="card-body row">
-              <div class="col-lg-2">
-              </div>
-              <form class="form-inline col-lg-10 search-form">
-                <input class="form-control form-control-lg search-field" type="search" placeholder="Search events here" aria-label="Search">
-                <button class="btn btn-secondary btn-large search-btn" type="submit">Search</button>
-              </form>
-            </div>
-          </div>
-          <div class="hashtags">
-            <ul>
-              <li>
-                #Odessa
-              </li>
-              <li>
-                #Music Festival
-              </li>
-              <li>
-                #Rock concert
-              </li>
-              <li>
-                #Night clubs
-              </li>
-              <li>
-                #Odessa
-              </li>
-              <li>
-                #Music Festival
-              </li>
-              <li>
-                #Rock concert
-              </li>
-            </ul>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
