--- afisha-client/src/app/search/search.component.scss
@@ -1,114 +1,0 @@
-.search-body {
-  width: 100%;
-  height: 100%;
-  background-image: linear-gradient(rgba(0,0,0,0.7), rgba(0,0,0,0.7)), url("../../assets/event-search.jpg");
-  background-size: cover;
-  background-position: center;
-}
-
-.clearfix {zoom: 1;}
-.clearfix:after {
-  content: '.';
-  clear: both;
-  display: block;
-  height: 0;
-  visibility: hidden;
-}
-
-/* Search container */
-
-.search-container {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-}
-
-.content {
-  text-align: center;
-}
-
-h1 {
-  color: #FCE4EC;
-}
-
-.card {
-  width: auto;
-  height: auto;
-}
-
-.search-form {
-  display: inline-block;
-  width: 100%;
-  text-align: center;
-}
-
-.search-field {
-  border: 0 transparent;
-  width: 80%;
-}
-
-.search-btn {
-  text-align: right;
-  height: 100%;
-}
-
-.search-btn :hover,
-.search-btn :active {
-  background-color: #EC407A;
-}
-
-.hashtags li {
-  background-color: #EC407A;
-  color: white;
-  font-size: 110%;
-  display: inline-block;
-  margin: 10px 10px 0 10px;
-  border-radius: 10%;
-  padding: 5px;
-}
-
-/* Search navi */
-.logo {
-  height: 100px;
-  width: auto;
-  float: left;
-  margin-top: 20px;
-}
-
-.clearfix::after {
-  content: "";
-  clear: both;
-  display: table;
-}
-
-.search-nav {
-  float: right;
-  list-style: none;
-  margin-top: 55px;
-}
-
-.search-nav li {
-  display: inline-block;
-  margin-left: 40px;
-}
-
-.search-nav li a:link,
-.search-nav li a:visited {
-  padding: 8px 0;
-  color: #FCE4EC;
-  text-decoration: none;
-  text-transform: uppercase;
-  font-size: 110%;
-  border-bottom: 2px solid transparent;
-  -webkit-transition: border-bottom 0.2s;
-  transition: border-bottom 0.2s;
-}
-
-.search-nav li a:hover,
-.search-nav li a:active {
-  border-bottom: 2px solid #EC407A;
-  color: #EC407A;
-}
