--- afisha-client/src/app/user/login.component.html
@@ -1,27 +1,0 @@
-<div class="container">
-  <div class="row justify-content-md-center">
-    <div class="col-3"></div>
-    <div class="col-4">
-      <mat-card class="card" role="group">
-        <mat-card-title>
-          <h3>Login</h3>
-        </mat-card-title>
-        <mat-card-content>
-          <form [formGroup]="loginForm" (ngSubmit)="login(loginForm.value)" autocomplete="off" novalidate>
-            <mat-form-field [ngClass]="{'invalid': !validateInputField('email')}">
-              <input matInput placeholder="email" type="email" formControlName="email">
-            </mat-form-field>
-            <mat-form-field [ngClass]="{'invalid': !validateInputField('password')}">
-              <input matInput placeholder="password" type="password" formControlName="password">
-            </mat-form-field>
-            <mat-card-actions>
-              <button mat-raised-button color="primary" [disabled]="!validateLoginForm()">Login</button>
-              <button mat-raised-button (click)="cancel()">Cancel</button>
-            </mat-card-actions>
-          </form>
-        </mat-card-content>
-      </mat-card>
-    </div>
-    <div class="col-3"></div>
-  </div>
-</div>
